// continental-frontend/src/interfaces/Api.interface.ts
interface Area {
  areaId: number;
  nombreGeneral: string;
  unidadOrganizativaSap: string;
}

interface Grupo {
  grupoId: number;
  rol: string;
  identificadorSAP: string;
  personasPorTurno: number;
  duracionDeturno: number;
}
export interface UsuarioInfoDto {
  id: number;
  fullName: string;
  username: string;
  unidadOrganizativaSap: string;
  rol: string;
  fechaIngreso: string;
  nomina: string;
  area: Area;
  grupo: Grupo;
}
// ...
export interface SolicitudReprogramacion {
  id: number;
  empleadoId: number;
  nombreEmpleado: string;
  nominaEmpleado: string;
  areaEmpleado: string;
  grupoEmpleado: string;
  vacacionOriginalId: number;
  fechaOriginal: string;
  fechaNueva: string;
  motivo: string;
  estadoSolicitud: 'Pendiente' | 'Aprobada' | 'Rechazada';
  requiereAprobacion: boolean;
  porcentajeCalculado: number;
  fechaSolicitud: string;
  solicitadoPor: string;
  fechaAprobacion: string | null;
  aprobadoPor: string | null;
  motivoRechazo: string | null;
  puedeAprobar: boolean;
}